Refactor ConstantNode to use long

+++ src/main/java/ca/wolfram/beta/symath/base/BaseNode.java
@@ -10,10 +10,10 @@ import ca.wolfram.beta.symath.MathNode;
 public class BaseNode {
 
     public static MathNode
-            ZERO = create(0),
-            ONE = create(1);
+            ZERO = create(0l),
+            ONE = create(1l);
 
-    public static MathNode create(int i) {
+    public static MathNode create(long i) {
         return new ConstantNode(i);
     }
